+++ src/provider/semanticTokensRangeManager.ts
@@ -2,7 +2,6 @@
 import { v4 as uuid } from 'uuid'
 import { CancellationToken, Disposable, DocumentSelector, Range, SemanticTokens, SemanticTokensLegend } from 'vscode-languageserver-protocol'
 import { TextDocument } from 'vscode-languageserver-textdocument'
-import { CancellationError } from '../util/errors'
 import { DocumentRangeSemanticTokensProvider } from './index'
 import Manager, { ProviderItem } from './manager'
 const logger = require('../util/logger')('semanticTokensRangeManager')
@@ -35,7 +34,7 @@ export default class SemanticTokensRangeManager extends Manager<DocumentRangeSem
     try {
       return await Promise.resolve(provider.provideDocumentRangeSemanticTokens(document, range, token))
     } catch (err) {
-      if (err instanceof CancellationError) return null
+      if (token.isCancellationRequested) return null
       throw err
     }
   }